requirements.txt
beautifulsoup4==4.13.1
Flask==3.1.0
loguru==0.7.3
matplotlib==3.10.0
pandas==2.2.3
pytest==8.3.4
PyYAML==6.0.2
Requests==2.32.3
schedule==1.2.2
